+++ src/api/user.api.ts
@@ -10,11 +10,12 @@ import {
   userIdSchema,
 } from "../schema.zod.js";
 import { auth } from "../lib/auth.js";
+import {authenticateAdmin, authenticate} from "../authentication/jwtauth.js"
 
 export const userApi = new Hono();
 
 //ná í
-userApi.get("/", zValidator("query", pagingSchema), async (c) => {
+userApi.get("/", authenticate,zValidator("query", pagingSchema), async (c) => {
   const limit = c.req.valid("query").limit;
   const offset = c.req.valid("query").offset;
 
@@ -35,7 +36,7 @@ userApi.get("/", zValidator("query", pagingSchema), async (c) => {
 });
 
 //Ná í eftir id eða slug
-userApi.get("/:id", zValidator("param", userIdSchema), async (c) => {
+userApi.get("/:id", authenticate,zValidator("param", userIdSchema), async (c) => {
   const id = c.req.valid("param").id;
 
   const user = await prisma.user.findUnique({ where: { id: id } });
@@ -49,7 +50,7 @@ userApi.get("/:id", zValidator("param", userIdSchema), async (c) => {
 
 //Búa til
 userApi.post(
-  "/",
+  "/",authenticateAdmin,
   zValidator("json", createUserSchema, (result, c) => {
     if (!result.success) {
       return c.json("Bad request", 400);
@@ -77,7 +78,7 @@ userApi.post(
 
 //Uppfæra
 userApi.put(
-  "/:id",
+  "/:id",authenticateAdmin,
   zValidator("json", updateUserSchema, (result, c) => {
     if (!result.success) {
       return c.json("Bad request", 400);
@@ -104,7 +105,7 @@ userApi.put(
 );
 
 //Eyða
-userApi.delete("/:id", zValidator("param", userIdSchema), async (c) => {
+userApi.delete("/:id", authenticateAdmin,zValidator("param", userIdSchema), async (c) => {
   const id = c.req.valid("param").id;
 
   await prisma.user.delete({
